requirements.txt
Django==1.9.5
Markdown==2.6.6
Pillow==3.2.0
django-filter==0.13.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.4.5
requests==2.9.1